requirements.txt
langchain
langgraph
langchain_community
langchain-groq
python-dotenv
streamlit
langchain-tavily
psycopg[binary,pool]
langgraph-checkpoint-postgres
FastAPI
-e .